refactor(trashbin): Move columns into separate file
